fix, refactor: animations, NBT, and networking
Big change, but essentially this fixed most of what was wrong with the musket logic. Networking was broken, animations were scuffed. I refactored a lot to get the code in a better state. I don't like how the NBT date updates every tick, but I don't know a better way.