update mod name in build.gradle
known issue: mixins do not work on the jar. I'm trying to find out why, but I have not luck so far. I hope this doesn't take too long to figure out. I think the issue is that the Mixin uses the Yarn mappings instead of the intermediary names, but I'm not sure? Looking at other mods, the mixins don't use intermediate names to specify where to Inject and such, so I'm not sure. At least it's only the mixin-related stuff that's not fully working. Everything else works, and I'm so happy about that. Hopefully this issue won't take too long to resolve.